Changes score types to decimal

+++ FMS.Domain/Dto/Status/StatusCreateDto.cs
@@ -79,7 +79,7 @@ namespace FMS.Domain.Dto
         public bool FinancialAssurance { get; set; }
 
         [Display(Name = "Cost Estimate")]
-        public double? CostEstimate { get; set; }
+        public decimal? CostEstimate { get; set; }
 
         [Display(Name = "Cost Estimate Date")]
         public DateOnly? CostEstimateDate { get; set; }
